[server] Script to update current season items
Added script to update all current season items. This
includes things such as: downloading data, re-generating
data, generating the HTML dashboard, and pushing to GitHub
Pages.

Script is meant to be attached to a cron job or scheduler,
or can be ran standalone if needed.

To set-up as a cron job (example on Raspberry Pi):
  - Run command: crontab -e
  - Add: PATH="/usr/local/bin:/usr/bin:/bin:/home/pi/.local/bin/"
    -> This ensures the cron job knows where to find uv
       (in this case: /home/pi/.local/bin/) and preserves default
       PATH variables
    -> This is needed because cron jobs run in a limited environment
  - Add entry to cron job to run this script
    -> Pipe outputs to file by appending: >> <outfile.log> 2>&1
    -> Use absolute paths in cron entry