[streamlit] Add requirements.txt
Right now, streamlit finds the pyproject.toml file in the root and
tries to install packages from it. However streamlit does not
currently support submodules, so it fails when trying to install
espn_league54078_fantasy_stats.

Added requirements.txt file in streamlit/ folder where the
app is deployed. Following this comment to see if this works:
https://discuss.streamlit.io/t/how-to-add-local-repository-code-to-deploy
/48172/4