[streamlit] Add initial current season page
Updated main page of streamlit app to display current
season stats. Start with a cumulative daily plot with
some stats on the side and option to filter for last
X days of the season.

Probably not the cleanest implementation so far since
streamlit re-runs the entire script top-to-bottom on
an update. Could find ways to improve the architecture
in the future.